refactor: give storyline handlers independent helper owners
