+++ src/evidenceforge/generation/engine/storyline_helpers/__init__.py
@@ -1,0 +1,4 @@
+# Copyright (c) 2026 Cisco Systems, Inc. and its affiliates
+# SPDX-License-Identifier: MIT
+
+"""Focused helper owners shared by storyline coordination and typed handlers."""
